Move infer_storage_options definition to dask.utils
Distributed imports `dask.utils.infer_storage_options`. Since we want
dask to work with older versions of distributed, we need to keep that
import around for a while. We can't define it in `dask.bytes.utils`
since that results in circular imports. As such, we keep the definition
in `dask.utils`, import it in `dask.bytes.utils`, and always use the
later internally. Once distributed's dask requirements have been bumped
we can move the definition.
